requirements.txt
alembic==1.13.1
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.0
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
filelock==3.13.1
Flask==3.0.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mlflow==2.9.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
optuna==3.5.0
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
protobuf==4.25.2
pyarrow==14.0.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.17.0
